Add eviction list to bucket cache
This fixes a bug that was causing partially flushed data to be loaded
by another thread before the data has been fully flushed out. The
eviction list is a linked list of buckets currently being flushed, and
threads wishing to load that bucket wait until the flushing thread
signals completion.